// web/src/api/generated/model/applicationStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Campus Portal API
 * OpenAPI spec version: 0.1.0
 */

export type ApplicationStatus = typeof ApplicationStatus[keyof typeof ApplicationStatus];


export const ApplicationStatus = {
  draft: 'draft',
  submitted: 'submitted',
  in_review: 'in_review',
  more_info: 'more_info',
  accepted: 'accepted',
  rejected: 'rejected',
  withdrawn: 'withdrawn',
} as const;